ape: make eight headers self-contained; sys/uio.h returns ssize_t
	/sys/include/ape/sys/uio.h:20 syntax error, last name: size_t

<sys/uio.h> used size_t for iov_len and included nothing that defines
it, so it compiled only where something else had got there first.
LibreSSL's include/compat/sys/uio.h reaches it through #include_next
before anything has. Fixed with <stddef.h> and <sys/types.h>.

While there: writev and readv were declared and defined returning int,
and POSIX gives both ssize_t. That is the readlinkat bug again -- the
note in ap/unistd/at_functions.c spells it out -- where on amd64 the
callee writes four bytes and an ssize_t caller reads eight, leaving
whatever was in the top half of the register. Header and both
definitions now say ssize_t. Not const-qualified, unlike POSIX's "const
struct iovec *": the definitions are not, and a declaration that
disagrees with its definition is worse than one that is merely lax.

Swept the rest of sys/include/ape for the same shape -- a header naming
size_t, ssize_t, time_t, va_list or FILE with nothing in its include
closure that defines them. 34 hits, of which eight are APE's own and are
fixed here: grp.h, pwd.h, langinfo.h, utime.h, sys/random.h, fts.h,
arpa/nameser.h and bio.h.

The other 26 are upstream copies -- curl/, lzma/, openssl/, png.h,
readline/, and the C++ and Objective-C runtime headers -- and are left
alone deliberately. None of them is self-contained upstream either;
curl/easy.h is meant to arrive through curl/curl.h and lzma/base.h
through lzma.h. Patching them would buy nothing and drift from the
originals, which for openssl/ and curl/ are copies this tree has to keep
in step by hand.

This class only surfaces now because LibreSSL's compat/ headers pull
APE's in earlier than anything has before.

+++ sys/include/ape/sys/uio.h
@@ -8,6 +8,17 @@ extern "C" {
 #pragma lib "/$M/lib/ape/libap.a"
 
 /*
+ * size_t for iov_len and ssize_t for the two return values. This header
+ * used both and included nothing, so it compiled only when something
+ * else had defined them first. LibreSSL's include/compat/sys/uio.h
+ * reaches this one through #include_next before anything else has:
+ *
+ *   /sys/include/ape/sys/uio.h:20 syntax error, last name: size_t
+ */
+#include <stddef.h>
+#include <sys/types.h>
+
+/*
  * Copyright (c) 1982, 1986 Regents of the University of California.
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
@@ -20,8 +31,20 @@ struct iovec {
       size_t iov_len;    /* was int */
   };
 
-extern int writev(int, struct iovec*, int);
-extern int readv(int, struct iovec*, int);
+/*
+ * POSIX gives both an ssize_t return, and they said int. On amd64 that
+ * is the readlinkat bug again (see ap/unistd/at_functions.c): the
+ * callee writes four bytes where an ssize_t caller reads eight, so the
+ * top half is whatever was in the register. ap/unistd/writev.c and
+ * readv.c return ssize_t to match.
+ *
+ * Not const-qualified, unlike POSIX's "const struct iovec *", because
+ * the definitions are not either and a declaration that disagrees with
+ * its definition is worse than one that is merely lax. A caller passing
+ * a non-const array is unaffected.
+ */
+extern ssize_t writev(int, struct iovec*, int);
+extern ssize_t readv(int, struct iovec*, int);
 
 #ifdef __cplusplus
 }
